EDU-294 refactor: 모듈 분할
