+++ app/main.py
@@ -4,21 +4,21 @@ from dotenv import load_dotenv
 from pydantic import BaseModel, Field
 from starlette.responses import JSONResponse
 from fastapi.encoders import jsonable_encoder
-from app.gpt_prompt import build_growth_feedback_prompt
-from app.mongo_feedback import save_feedback_cache
+from app.utils.gpt_prompt import build_growth_feedback_prompt
+from app.services.mongo_feedback import save_feedback_cache
 import openai
 import os
 from datetime import date
 from typing import List, Dict
-from app.mongodb import db
+from app.configs.mongodb import db
 
 
 #  라우터들 먼저 import
-from app.routes.assessment import router as assessment_router
-from app.routes.post_assessment import router as post_assessment_router
-from app.routes.activity_log import router as activity_log_router
-from app.routes.roadmap_route import router as roadmap_router
-from app.routes.feedback_route import router as feedback_router
+from app.routers.endpoints.assessment import router as assessment_router
+from app.routers.endpoints.post_assessment import router as post_assessment_router
+from app.routers.endpoints.activity_log import router as activity_log_router
+from app.routers.roadmap_route import router as roadmap_router
+from app.routers.feedback_route import router as feedback_router
 
 #  환경 변수 및 OpenAI 키 설정
 load_dotenv()
