+++ app/routers/roadmap_route.py
@@ -1,9 +1,9 @@
 from fastapi import APIRouter
 from pydantic import BaseModel
-from app.mongodb import db
-from app.mongo_roadmap import save_roadmap_cache
-from app.gpt_prompt import build_roadmap_prompt
-from app.roadmap_rag import generate_roadmap_rag
+from app.configs.mongodb import db
+from app.services.mongo_roadmap import save_roadmap_cache
+from app.utils.gpt_prompt import build_roadmap_prompt
+from app.services.roadmap_rag import generate_roadmap_rag
 
 router = APIRouter()
