+++ app/services/mongo_roadmap.py
@@ -1,6 +1,6 @@
 # app/mongo_roadmap.py
 
-from app.mongodb import db
+from app.configs.mongodb import db
 from datetime import datetime
 
 async def save_roadmap_cache(user_id, roadmap):
